Parse fractional treasury and amenity figures instead of dropping them
GetGoldBalance/GetGoldYield/food surplus are floats in a real game (verified
live: 428.8125 / 55.953125), but the maintenance and amenities parsers used a
bare int() derived from a single early-game observation where gold happened
to be whole. The fraction raised ValueError, was swallowed, and the client
told the player the game's reply "could not be read" -- a false cause; the
reply was perfect.

- queries.py: shared _parse_number() accepts both int and float text,
  preserving int-ness when there is no decimal point; used by the
  maintenance and amenities parsers.
- base.py: Maintenance.gold/gold_yield and CityAmenities.food_surplus are now
  float; net_gold and unattributed follow without forcing a round.
- client.py: a genuinely unparseable field now blames "the advisor could not
  read" the figures, never "the reply could not be read".
- evidence.py/app.js/briefing.js: floats round for DISPLAY only (prose notes,
  table cells); stored/serialized values keep full precision.
- Added a regression test asserting a normal capture() asks exactly three
  tuner queries and an acting one asks four.

+++ civ_advisor/tuner/base.py
@@ -82,7 +82,11 @@ class CityAmenities:
     from_civics: int
     from_entertainment: int
     housing: int
-    food_surplus: int
+    # A float, not an int: amenity counts and housing are verified integers in a
+    # real mid-game state, but food surplus is fractional in Civ VI generally --
+    # the same fact that makes Maintenance.gold and .gold_yield floats below, and
+    # the same live probe that caught it (see queries.py's `_parse_number`).
+    food_surplus: float
 
     @property
     def unexplained(self) -> int:
@@ -106,8 +110,14 @@ class Maintenance:
     buildings: int
     districts: int
     units: int
-    gold: int
-    gold_yield: int
+    # Floats, not ints: GetGoldBalance and GetGoldYield answered 428.8125 and
+    # 55.953125 against a real live treasury on 2026-09-13 (turn well past the
+    # game's start). The original `int(...)` assumption came from a single
+    # early-game reading where gold happened to be a whole number (152) -- one
+    # observation standing in for a validator. queries.py's `_parse_number`
+    # documents the full story.
+    gold: float
+    gold_yield: float
 
     @property
     def unattributed(self) -> int:
